- feature: Added and adapted changes did by Nico
