+++ visor-extension/package.json
@@ -24,7 +24,6 @@
     "@types/react-dom": "^17.0.1",
     "@types/react-router": "^5.1.17",
     "@types/react-router-dom": "^5.3.1",
-    "@types/uuid": "^9.0.1",
     "babel-loader": "^8.2.2",
     "copy-webpack-plugin": "^9.0.1",
     "css-loader": "^6.4.0",
@@ -58,11 +57,10 @@
     "rxjs": "^7.5.2",
     "ts-node": "^10.3.0",
     "urbit-ob": "^5.0.1",
-    "uuid": "^9.0.0",
     "zustand": "^3.6.9"
   },
   "engines": {
-    "node": "14",
-    "npm": "6"
+    "node": "14.17.6",
+    "npm": "6.14.15"
   }
 }
