+++ visor-extension/src/components/modal/commands/History.tsx
@@ -1,7 +1,6 @@
 import { Command } from '../types';
 import React from 'react';
 import AgrihanInterface from '@urbit/http-api';
-import { addDmMessage } from '@urbit/api';
 import dmIcon from '../../../icons/history.svg';
 
 const Icon = () => <img src={dmIcon} />;
