+++ visor-extension/src/custom.d.ts
@@ -2,3 +2,5 @@ declare module '*.svg' {
   const content: any;
   export default content;
 }
+
+declare module '@urbit/api';
