+++ visor-extension/src/index.d.ts
@@ -1,2 +1,1 @@
 declare module '*.png';
-declare module '@urbit/api';
